Add lastRecordTime, delay, isRecording variables to prevent gesture from being recognised immediately after recording
isRecording is set to true when Save() function starts, and false when it ends. lastRecordTime is also set in Save() function to keep track of when the last gesture was recorded. In Recognize(), there is now a check for if gesture is not currently being recorded, and that one second (delay) has passed from when gesture was recorded.