Updated Gesture constructor and readGesturesFromJSON method.
This was done to better organize fingerData to be mapped to the hand model. readJSON creates a new Gesture with the retrieved data (should prevent gestures dictionary from being null). Data is organised so that each finger is identified by a unique name (e.g., thumb, index, middle) as the key, and the associated list of Vector3 values represents the finger positions at different points in time, allowing for easier mapping to hand model. JArray allows us to easily iterate over elements in the JSON array.