Set default position/rotations of hand model at Start
Issue with reference/default position of the hand model becoming the previously played gesture. Setting it once at the start ensures all gestures are played based on actual starting point of the hand model. Also UpdateNestedFingerBones() was updated to properly access and update the finger bones (not the root finger).