Add checks to determine why positional/rotational calculations are incorrect in UpdateNestedFingerBones()
I seems like there is a difference between what we get from saving the reference position/rotation in the hand model and what the hand models actual default position/rotation is. For the most part it looks like the json saved positions and rotations match what we're getting when we access those bones.