Update Recognise function to recognise whole hand rotations in Gestures
Changed detectionThreshold for whole hand rotations in CompareRotationData (50f worked well). Changed way we get the rotation for the frame (Quaternion motionHandRotation = motionFrameData["HandPosition"].rotation;). This gets the correct rotation for the frame (matches the saved json). Calculated delta rotation to account for the relative rotation between the initial hand position in the motion frame data and the current hand position